Lchoy/website migration (#391)
This commit sets up the hidivelab.org website which is sourced from an S3 bucket and served via a CloudFront distribution.  The static-site.yaml is a Cloudformation template for deploying the AWS resources. It provisions two S3 buckets for hosting the content and logs, respectively. A certificate is created to support https. The CloudFront distribution includes a function association to add index.html to request URLs that do not contain a file name. This does not include redirection of the existing gehlenborglab.org site to hidivelab.org.

Updates GitHub actions to deploy to S3 bucket in hdv_lab account.

* Upgrade to aws-actions/configure-aws-credentials@v3

* Add comment and reference for CloudFront function.